NotesApp/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Columnar==1.4.1
dictionary==1.0
distlib==0.3.4
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.5.1
gunicorn==20.1.0
idna==3.3
lxml==4.7.1
numpy==1.22.1
pandas==1.4.0
platformdirs==2.5.0
Py-Dictionary==4.1.4
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
string-color==1.2.1
toolz==0.11.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.13.1
wcwidth==0.2.5
whitenoise==6.0.0
